requirements.txt
boto3==1.33.10
Jinja2==3.1.2
glueops-helpers @ https://github.com/GlueOps/python-glueops-helpers-library/archive/refs/tags/v0.2.0.zip